add bottomNavigationView
