+++ app/build.gradle
@@ -29,6 +29,10 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
+
+    dataBinding{
+        enabled = true
+    }
 }
 
 dependencies {
