+++ app/src/main/res/layout/fragment_message.xml
@@ -2,12 +2,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.navigation.HomeFragment">
+    tools:context="com.example.navigation.MessageFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="HomeFragment" />
+        android:text="MessageFragment" />
 
 </FrameLayout>
