+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -4,14 +4,18 @@
 
     <item
         android:id="@+id/homeFragment"
+        android:icon="@drawable/home_48px"
         android:title="Home" />
     <item
         android:id="@+id/searchFragment"
+        android:icon="@drawable/search_48px"
         android:title="Search" />
     <item
-        android:id="@+id/NotificationMenu"
+        android:id="@+id/notificationFragment"
+        android:icon="@drawable/notifications_48px"
         android:title="Notification" />
     <item
-        android:id="@+id/Menu"
+        android:id="@+id/messageFragment"
+        android:icon="@drawable/mail_48px"
         android:title="DirectMessage" />
 </menu>
