+++ app/src/main/res/navigation/navigation.xml
@@ -9,10 +9,22 @@
         android:id="@+id/homeFragment"
         android:name="com.example.navigation.HomeFragment"
         android:label="placeholder_layout"
-        tools:layout="@layout/fragment_home" />
+        tools:layout="@layout/fragment_home" >
+    </fragment>
     <fragment
         android:id="@+id/searchFragment"
         android:name="com.example.navigation.SearchFragment"
         android:label="fragment_search"
-        tools:layout="@layout/fragment_search" />
+        tools:layout="@layout/fragment_search">
+    </fragment>
+    <fragment
+        android:id="@+id/messageFragment"
+        android:name="com.example.navigation.MessageFragment"
+        android:label="fragment_message"
+        tools:layout="@layout/fragment_message" />
+    <fragment
+        android:id="@+id/notificationFragment"
+        android:name="com.example.navigation.NotificationFragment"
+        android:label="fragment_notification"
+        tools:layout="@layout/fragment_notification" />
 </navigation>
